backend/requirements.txt
django==3.2.18
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
pillow==9.4.0
python-dotenv==0.21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
